fix: multiple fixes according to review
