+++ .dependabot/config.yml
@@ -36,14 +36,14 @@ update_configs:
           dependency_type: "development"
           update_type: "all"
   - package_manager: "javascript"
-    directory: "/templates/vue3/packagejsons/assemblyscript"
+    directory: "/templates/vue/packagejsons/assemblyscript"
     update_schedule: "live"
     automerged_updates:
       - match:
           dependency_type: "development"
           update_type: "all"
   - package_manager: "javascript"
-    directory: "/templates/vue3/packagejsons/rust"
+    directory: "/templates/vue/packagejsons/rust"
     update_schedule: "live"
     automerged_updates:
       - match:
