+++ .gitignore
@@ -1,12 +1,9 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
 node_modules
 /.pnp
 .pnp.js
-/out
-/tmp-project
 out
 package-lock.json
 /neardev
