+++ templates/vue/as-pect.config.js
