+++ templates/vue/asconfig.js
