+++ templates/vue/assembly
