+++ templates/vue/babel.config.js
