+++ templates/vue/copy-dev-account.js
