+++ templates/vue/near.gitignore
