+++ templates/vue/package.json
