+++ templates/vue/packagejsons/assemblyscript/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-vue-project",
+  "name": "near-blank-project",
   "version": "0.1.0",
   "private": true,
   "scripts": {
