+++ templates/vue/packagejsons/rust/package.json
