+++ templates/vue/src/assets
