+++ templates/vue/src/components/Notification.vue
