+++ templates/vue/src/config.js
