+++ templates/vue/src/favicon.ico
