+++ templates/vue/src/global.css
