+++ templates/vue/src/jest.init.js
@@ -1,0 +1,1 @@
+import 'regenerator-runtime'
