+++ templates/vue/src/main.js
