--- templates/vue3/src/index.html
@@ -1,16 +1,0 @@
-﻿<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="./assets/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="apple-touch-icon" href="./assets/favicon.ico" />
-    <title>Welcome to NEAR with Vue</title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="app"></div>
-    <!-- built files will be auto injected -->
-  <script src="./main.js"></script>
-</body>
-</html>
