--- templates/vue3/vue.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-    chainWebpack: config => {
-        config
-            .plugin('html')
-            .tap(args => {
-                args[0].title = "My NEAR Vue App";
-                return args;
-            })
-    }
-}
