+++ test/test-new-project.js
@@ -13,10 +13,10 @@ if (![undefined, 'assemblyscript', 'rust'].includes(process.env.CONTRACT)) {
   process.exit(1)
 }
 
-if (![undefined, 'react', 'vanilla','vue3'].includes(process.env.FRONTEND)) {
+if (![undefined, 'react', 'vanilla','vue'].includes(process.env.FRONTEND)) {
   console.error(
     `Unknown FRONTEND template '${process.env.FRONTEND}'; ` +
-        'use \'react\' or \'vanilla\' or \'vue3\''
+        'use \'react\' or \'vanilla\' or \'vue\''
   )
   process.exit(1)
 }
@@ -27,7 +27,7 @@ const contracts = process.env.CONTRACT
 
 const frontends = process.env.FRONTEND
   ? [process.env.FRONTEND]
-  : ['vanilla', 'react', 'vue3']
+  : ['vanilla', 'react', 'vue']
 
 const commands = contracts.map(c => frontends.map(f => (
   `node index.js tmp-project --contract=${c} --frontend=${f}`
